Dockerfiles/API_Python/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
cassandra-driver==3.25.0
Flask-Cors==3.0.10
